refactor: lurchThumbnailを改修
